feat(app): separated all the router routes into different files
