+++ packages/app/src/Domains/Auth/index.ts
@@ -1,2 +1,6 @@
+export * from './Auth.router';
 export * from './Auth.routes';
 export * from './Auth.service';
+export * from './Components';
+export * from './Hooks';
+export * from './Pages';
