+++ packages/app/src/Domains/Main/Main.routes.tsx
@@ -1,8 +1,1 @@
-import { Route } from 'react-router-dom';
-import { MainPage } from './Pages/Main.page';
-
 export const MAIN_ROUTE = '/main';
-
-export const MainRouter = [
-  <Route key="main" path={MAIN_ROUTE} element={<MainPage />} />,
-];
