+++ packages/app/src/Domains/Main/index.ts
@@ -1,1 +1,2 @@
+export * from './Main.router';
 export * from './Main.routes';
