+++ packages/app/src/Domains/Users/index.ts
@@ -2,5 +2,6 @@ export * from './Components';
 export * from './Hooks';
 export * from './Pages';
 export * from './User.entity';
-export * from './Users.service';
+export * from './Users.router';
 export * from './Users.routes';
+export * from './Users.service';
